refactor: estrai domain/profile.py (csv_preview) da cli/inspect/
csv_preview in domain/profile.py. Comando Typer + run_profile restano in cli/.
Aggiorna MCP a importare da domain/.
Azzera dipendenza MCP→cli/ per profile_ops.
